started code cleanup, trying a fix to aerodynamics_super.py to allow for different propulsors
