--- regression/scripts/aerodynamics/mission_B737.py
@@ -1,1139 +1,0 @@
-# full_setup.py
-# 
-# Created:  SUave Team, Aug 2014
-# Modified: 
-
-""" setup file for a mission with a 737
-"""
-
-
-# ----------------------------------------------------------------------
-#   Imports
-# ----------------------------------------------------------------------
-
-import SUAVE
-from SUAVE.Core import Units
-
-import numpy as np
-import pylab as plt
-
-import copy, time
-
-from SUAVE.Core import (
-Data, Container,
-)
-
-from SUAVE.Methods.Propulsion.turbofan_sizing import turbofan_sizing
-from SUAVE.Input_Output.Results import  print_parasite_drag,  \
-                                        print_compress_drag, \
-                                        print_engine_data,   \
-                                        print_mission_breakdown, \
-                                        print_weight_breakdown
-# ----------------------------------------------------------------------
-#   Main
-# ----------------------------------------------------------------------
-
-def main():
-    
-    configs, analyses = full_setup()
-    
-    simple_sizing(configs)
-    
-    configs.finalize()
-    analyses.finalize()
-    
-    # weight analysis
-    weights = analyses.configs.base.weights
-    breakdown = weights.evaluate()      
-    
-    # mission analysis
-    mission = analyses.missions.base
-    results = mission.evaluate()
-
-    # print weight breakdown
-    print_weight_breakdown(configs.base,filename = 'weight_breakdown.dat')
-
-    # print engine data into file
-    print_engine_data(configs.base,filename = 'B737_engine_data.dat')
-
-    # print parasite drag data into file
-	# define reference condition for parasite drag
-    ref_condition = Data()
-    ref_condition.mach_number = 0.3
-    ref_condition.reynolds_number = 12e6     
-    print_parasite_drag(ref_condition,configs.cruise,analyses,'B737_parasite_drag.dat')
-    
-    # print compressibility drag data into file
-    print_compress_drag(configs.cruise,analyses,filename = 'B737_compress_drag.dat')
-    
-    # print mission breakdown
-    print_mission_breakdown(results,filename='B737_mission_breakdown.dat')
-    
-    # load older results
-    #save_results(results)
-    old_results = load_results()   
-    
-    # plt the old results
-    plot_mission(results)
-    plot_mission(old_results,'k-')
-    
-    # check the results
-    check_results(results,old_results)
-    
-##	# print some results, for check agaist Aviation paper
-##    end_segment = results.segments[-1]
-##    time     = end_segment.conditions.frames.inertial.time[-1,0] / Units.min
-##    distance = end_segment.conditions.frames.inertial.position_vector[-1,0] / Units.nautical_miles
-##    mass_end     = end_segment.conditions.weights.total_mass[-1,0]
-##    mass_begin   = results.segments[0].conditions.weights.total_mass[0,0]
-##    fuelburn = (mass_begin- mass_end) / Units.lbs
-##    
-##    print 'Time: {:5.0f} min , Distance: {:5.0f} nm ; FuelBurn: {:5.0f} lbs'.format(time,distance,fuelburn)
-
-    return
-
-
-# ----------------------------------------------------------------------
-#   Analysis Setup
-# ----------------------------------------------------------------------
-
-def full_setup():
-    
-    # vehicle data
-    vehicle  = vehicle_setup()
-    configs  = configs_setup(vehicle)
-    
-    # vehicle analyses
-    configs_analyses = analyses_setup(configs)
-    
-    # mission analyses
-    mission  = mission_setup(configs_analyses)
-    missions_analyses = missions_setup(mission)
-
-    analyses = SUAVE.Analyses.Analysis.Container()
-    analyses.configs  = configs_analyses
-    analyses.missions = missions_analyses
-    
-    return configs, analyses
-
-# ----------------------------------------------------------------------
-#   Define the Vehicle Analyses
-# ----------------------------------------------------------------------
-
-def analyses_setup(configs):
-    
-    analyses = SUAVE.Analyses.Analysis.Container()
-    
-    # build a base analysis for each config
-    for tag,config in configs.items():
-        analysis = base_analysis(config)
-        analyses[tag] = analysis
-        
-    # adjust analyses for configs
-    
-    # takeoff_analysis
-    analyses.takeoff.aerodynamics.settings.drag_coefficient_increment = 0.0000
-    
-    # landing analysis
-    aerodynamics = analyses.landing.aerodynamics
-    # do something here eventually
-    
-    return analyses
-
-def base_analysis(vehicle):
-
-    # ------------------------------------------------------------------
-    #   Initialize the Analyses
-    # ------------------------------------------------------------------     
-    analyses = SUAVE.Analyses.Vehicle()
-    
-    # ------------------------------------------------------------------
-    #  Basic Geometry Relations
-    sizing = SUAVE.Analyses.Sizing.Sizing()
-    sizing.features.vehicle = vehicle
-    analyses.append(sizing)
-    
-    # ------------------------------------------------------------------
-    #  Weights
-    weights = SUAVE.Analyses.Weights.Weights()
-    weights.vehicle = vehicle
-    analyses.append(weights)
-    
-    # ------------------------------------------------------------------
-    #  Aerodynamics Analysis
-    aerodynamics = SUAVE.Analyses.Aerodynamics.Fidelity_Zero()
-    aerodynamics.geometry = vehicle
-    
-    ## modify inviscid wings - linear lift model
-    #inviscid_wings = SUAVE.Analyses.Aerodynamics.Linear_Lift()
-    #inviscid_wings.settings.slope_correction_coefficient = 1.04
-    #inviscid_wings.settings.zero_lift_coefficient = 2.*np.pi* 3.1 * Units.deg    
-    #aerodynamics.process.compute.lift.inviscid_wings = inviscid_wings        
-    
-    ## modify inviscid wings - avl model
-    #inviscid_wings = SUAVE.Analyses.Aerodynamics.Surrogates.AVL()
-    #inviscid_wings.geometry = vehicle
-    #aerodynamics.process.compute.lift.inviscid_wings = inviscid_wings
-    
-    aerodynamics.settings.drag_coefficient_increment = 0.0000
-    analyses.append(aerodynamics)
-    
-    # ------------------------------------------------------------------
-    #  Stability Analysis
-    stability = SUAVE.Analyses.Stability.Fidelity_Zero()
-    stability.geometry = vehicle
-    analyses.append(stability)
-    
-    # ------------------------------------------------------------------
-    #  Energy
-    energy= SUAVE.Analyses.Energy.Energy()
-    energy.network = vehicle.propulsors #what is called throughout the mission (at every time step))
-    analyses.append(energy)
-    
-    # ------------------------------------------------------------------
-    #  Planet Analysis
-    planet = SUAVE.Analyses.Planets.Planet()
-    analyses.append(planet)
-    
-    # ------------------------------------------------------------------
-    #  Atmosphere Analysis
-    atmosphere = SUAVE.Analyses.Atmospheric.US_Standard_1976()
-    atmosphere.features.planet = planet.features
-    analyses.append(atmosphere)   
-    
-    # done!
-    return analyses    
-
-# ----------------------------------------------------------------------
-#   Define the Vehicle
-# ----------------------------------------------------------------------
-
-def vehicle_setup():
-    
-    # ------------------------------------------------------------------
-    #   Initialize the Vehicle
-    # ------------------------------------------------------------------    
-    
-    vehicle = SUAVE.Vehicle()
-    vehicle.tag = 'Boeing_737800'    
-    
-    
-    # ------------------------------------------------------------------
-    #   Vehicle-level Properties
-    # ------------------------------------------------------------------    
-
-    # mass properties
-    vehicle.mass_properties.max_takeoff               = 79015.8   # kg
-    vehicle.mass_properties.operating_empty           = 62746.4   # kg
-    vehicle.mass_properties.takeoff                   = 79015.8   # kg
-    vehicle.mass_properties.max_zero_fuel             = 0.9 * vehicle.mass_properties.max_takeoff
-    vehicle.mass_properties.cargo                     = 10000.  * Units.kilogram   
-    
-    vehicle.mass_properties.center_of_gravity         = [60 * Units.feet, 0, 0]  # Not correct
-    vehicle.mass_properties.moments_of_inertia.tensor = [[10 ** 5, 0, 0],[0, 10 ** 6, 0,],[0,0, 10 ** 7]] # Not Correct
-    
-    # envelope properties
-    vehicle.envelope.ultimate_load = 2.5
-    vehicle.envelope.limit_load    = 1.5
-
-    # basic parameters
-    vehicle.reference_area         = 124.862       
-    vehicle.passengers             = 170
-    vehicle.systems.control        = "fully powered" 
-    vehicle.systems.accessories    = "medium range"
-    
-    
-    # ------------------------------------------------------------------        
-    #   Main Wing
-    # ------------------------------------------------------------------        
-    
-    wing = SUAVE.Components.Wings.Main_Wing()
-    wing.tag = 'main_wing'
-    
-    wing.aspect_ratio            = 10.18
-    wing.sweeps.quarter_chord    = 25 * Units.deg
-    wing.thickness_to_chord      = 0.1
-    wing.taper                   = 0.16
-    wing.span_efficiency         = 0.9
-    
-    wing.spans.projected         = 35.66    
-    
-    wing.chords.root             = 6.81
-    wing.chords.tip              = 1.09
-    wing.chords.mean_aerodynamic = 4.235
-    
-    wing.areas.reference         = 124.862 
-    
-    wing.twists.root             = 4.0 * Units.degrees
-    wing.twists.tip              = -4.0 * Units.degrees
-    
-    wing.origin                  = [20,0,0]
-    wing.aerodynamic_center      = [3,0,0] 
-    
-    wing.vertical                = False
-    wing.symmetric               = True
-    wing.high_lift               = True
-    
-    wing.dynamic_pressure_ratio  = 1.0
-    
-    # add to vehicle
-    vehicle.append_component(wing)
-
-
-    # ------------------------------------------------------------------        
-    #  Horizontal Stabilizer
-    # ------------------------------------------------------------------        
-    
-    wing = SUAVE.Components.Wings.Wing()
-    wing.tag = 'horizontal_stabilizer'
-    
-    wing.aspect_ratio            = 6.16      #
-    wing.sweeps.quarter_chord    = 30 * Units.deg
-    wing.thickness_to_chord      = 0.08
-    wing.taper                   = 0.4
-    wing.span_efficiency         = 0.9
-    
-    wing.spans.projected         = 14.146      #
-
-    wing.chords.root             = 3.28
-    wing.chords.tip              = 1.31    
-    wing.chords.mean_aerodynamic = 8.0
-
-    wing.areas.reference         = 32.488    #
-    
-    wing.twists.root             = 3.0 * Units.degrees
-    wing.twists.tip              = 3.0 * Units.degrees  
-    
-    wing.origin                  = [50,0,0]
-    wing.aerodynamic_center      = [2,0,0]
-    
-    wing.vertical                = False 
-    wing.symmetric               = True
-    
-    wing.dynamic_pressure_ratio  = 0.9  
-    
-    # add to vehicle
-    vehicle.append_component(wing)
-    
-    
-    # ------------------------------------------------------------------
-    #   Vertical Stabilizer
-    # ------------------------------------------------------------------
-    
-    wing = SUAVE.Components.Wings.Wing()
-    wing.tag = 'vertical_stabilizer'    
-    
-    wing.aspect_ratio            = 1.91      #
-    wing.sweeps.quarter_chord    = 25 * Units.deg
-    wing.thickness_to_chord      = 0.08
-    wing.taper                   = 0.25
-    wing.span_efficiency         = 0.9
-    
-    wing.spans.projected         = 7.877      #    
-
-    wing.chords.root             = 6.60
-    wing.chords.tip              = 1.65
-    wing.chords.mean_aerodynamic = 8.0
-    
-    wing.areas.reference         = 32.488    #
-    
-    wing.twists.root             = 0.0 * Units.degrees
-    wing.twists.tip              = 0.0 * Units.degrees  
-    
-    wing.origin                  = [50,0,0]
-    wing.aerodynamic_center      = [2,0,0]    
-    
-    wing.vertical                = True 
-    wing.symmetric               = False
-    wing.t_tail                  = False
-    
-    wing.dynamic_pressure_ratio  = 1.0
-        
-    # add to vehicle
-    vehicle.append_component(wing)
-
-
-    # ------------------------------------------------------------------
-    #  Fuselage
-    # ------------------------------------------------------------------
-    
-    fuselage = SUAVE.Components.Fuselages.Fuselage()
-    fuselage.tag = 'fuselage'
-    
-    #fuselage.number_coach_seats    = vehicle.passengers
-    fuselage.seats_abreast         = 6
-    fuselage.seat_pitch            = 1
-    
-    fuselage.fineness.nose         = 1.6
-    fuselage.fineness.tail         = 2.
-    
-    fuselage.lengths.nose          = 6.4
-    fuselage.lengths.tail          = 8.0
-    fuselage.lengths.cabin         = 28.85 #44.0
-    fuselage.lengths.total         = 38.02 #58.4
-    fuselage.lengths.fore_space    = 6.
-    fuselage.lengths.aft_space     = 5.    
-    
-    fuselage.width                 = 3.74 #4.
-    
-    fuselage.heights.maximum       = 3.74  #4.    #
-    fuselage.heights.at_quarter_length          = 4. # Not correct
-    fuselage.heights.at_three_quarters_length   = 4. # Not correct
-    fuselage.heights.at_wing_root_quarter_chord = 4. # Not correct
-
-    fuselage.areas.side_projected  = 3.74* 38.02 #4.* 59.8 #  Not correct
-    fuselage.areas.wetted          = 446.718 #688.64
-    fuselage.areas.front_projected = 12.57
-    
-    fuselage.effective_diameter    = 3.74 #4.0
-    
-    fuselage.differential_pressure = 5.0e4 * Units.pascal # Maximum differential pressure
-    
-    # add to vehicle
-    vehicle.append_component(fuselage)
-    
-        
-    # ------------------------------------------------------------------
-    #   Turbofan Network
-    # ------------------------------------------------------------------    
-    
-    #instantiate the gas turbine network
-    turbofan = SUAVE.Components.Energy.Networks.Turbofan()
-    turbofan.tag = 'turbofan'
-    
-    # setup
-    turbofan.number_of_engines = 2.0
-    turbofan.bypass_ratio      = 5.4
-    turbofan.engine_length     = 2.71
-    turbofan.nacelle_diameter  = 2.05
-    
-    #compute engine areas
-    Awet    = 1.1*np.pi*turbofan.nacelle_diameter*turbofan.engine_length 
-    
-    #Assign engine areas
-    turbofan.areas.wetted  = Awet
-    
-    
-    
-    # working fluid
-    turbofan.working_fluid = SUAVE.Attributes.Gases.Air()
-    
-    
-    # ------------------------------------------------------------------
-    #   Component 1 - Ram
-    
-    # to convert freestream static to stagnation quantities
-    
-    # instantiate
-    ram = SUAVE.Components.Energy.Converters.Ram()
-    ram.tag = 'ram'
-    
-    # add to the network
-    turbofan.append(ram)
-
-
-    # ------------------------------------------------------------------
-    #  Component 2 - Inlet Nozzle
-    
-    # instantiate
-    inlet_nozzle = SUAVE.Components.Energy.Converters.Compression_Nozzle()
-    inlet_nozzle.tag = 'inlet_nozzle'
-    
-    # setup
-    inlet_nozzle.polytropic_efficiency = 0.98
-    inlet_nozzle.pressure_ratio        = 0.98
-    
-    # add to network
-    turbofan.append(inlet_nozzle)
-    
-    
-    # ------------------------------------------------------------------
-    #  Component 3 - Low Pressure Compressor
-    
-    # instantiate 
-    compressor = SUAVE.Components.Energy.Converters.Compressor()    
-    compressor.tag = 'low_pressure_compressor'
-
-    # setup
-    compressor.polytropic_efficiency = 0.91
-    compressor.pressure_ratio        = 1.14    
-    
-    # add to network
-    turbofan.append(compressor)
-
-    
-    # ------------------------------------------------------------------
-    #  Component 4 - High Pressure Compressor
-    
-    # instantiate
-    compressor = SUAVE.Components.Energy.Converters.Compressor()    
-    compressor.tag = 'high_pressure_compressor'
-    
-    # setup
-    compressor.polytropic_efficiency = 0.91
-    compressor.pressure_ratio        = 13.415    
-    
-    # add to network
-    turbofan.append(compressor)
-
-
-    # ------------------------------------------------------------------
-    #  Component 5 - Low Pressure Turbine
-    
-    # instantiate
-    turbine = SUAVE.Components.Energy.Converters.Turbine()   
-    turbine.tag='low_pressure_turbine'
-    
-    # setup
-    turbine.mechanical_efficiency = 0.99
-    turbine.polytropic_efficiency = 0.93     
-    
-    # add to network
-    turbofan.append(turbine)
-    
-      
-    # ------------------------------------------------------------------
-    #  Component 6 - High Pressure Turbine
-    
-    # instantiate
-    turbine = SUAVE.Components.Energy.Converters.Turbine()   
-    turbine.tag='high_pressure_turbine'
-
-    # setup
-    turbine.mechanical_efficiency = 0.99
-    turbine.polytropic_efficiency = 0.93     
-    
-    # add to network
-    turbofan.append(turbine)
-      
-    
-    # ------------------------------------------------------------------
-    #  Component 7 - Combustor
-    
-    # instantiate    
-    combustor = SUAVE.Components.Energy.Converters.Combustor()   
-    combustor.tag = 'combustor'
-    
-    # setup
-    combustor.efficiency                = 0.99 
-    combustor.alphac                    = 1.0     
-    combustor.turbine_inlet_temperature = 1450
-    combustor.pressure_ratio            = 0.95
-    combustor.fuel_data                 = SUAVE.Attributes.Propellants.Jet_A()    
-    
-    # add to network
-    turbofan.append(combustor)
-
-    
-    # ------------------------------------------------------------------
-    #  Component 8 - Core Nozzle
-    
-    # instantiate
-    nozzle = SUAVE.Components.Energy.Converters.Expansion_Nozzle()   
-    nozzle.tag = 'core_nozzle'
-    
-    # setup
-    nozzle.polytropic_efficiency = 0.95
-    nozzle.pressure_ratio        = 0.99    
-    
-    # add to network
-    turbofan.append(nozzle)
-
-
-    # ------------------------------------------------------------------
-    #  Component 9 - Fan Nozzle
-    
-    # instantiate
-    nozzle = SUAVE.Components.Energy.Converters.Expansion_Nozzle()   
-    nozzle.tag = 'fan_nozzle'
-
-    # setup
-    nozzle.polytropic_efficiency = 0.95
-    nozzle.pressure_ratio        = 0.99    
-    
-    # add to network
-    turbofan.append(nozzle)
-    
-    
-    # ------------------------------------------------------------------
-    #  Component 10 - Fan
-    
-    # instantiate
-    fan = SUAVE.Components.Energy.Converters.Fan()   
-    fan.tag = 'fan'
-
-    # setup
-    fan.polytropic_efficiency = 0.93
-    fan.pressure_ratio        = 1.7    
-    
-    # add to network
-    turbofan.append(fan)
-    
-    
-    # ------------------------------------------------------------------
-    #Component 10 : thrust (to compute the thrust)
-    thrust = SUAVE.Components.Energy.Processes.Thrust()       
-    thrust.tag ='compute_thrust'
- 
-    #total design thrust (includes all the engines)
-    thrust.total_design             = 2*24000. * Units.N #Newtons
- 
-    #design sizing conditions
-    altitude      = 35000.0*Units.ft
-    mach_number   = 0.78 
-    isa_deviation = 0.
-    
-    # add to network
-    turbofan.thrust = thrust
-
-    #size the turbofan
-    turbofan_sizing(turbofan,mach_number,altitude)   
-    
-    # add  gas turbine network turbofan to the vehicle 
-    vehicle.append_component(turbofan)      
-    
-    
-    # ------------------------------------------------------------------
-    #   Vehicle Definition Complete
-    # ------------------------------------------------------------------
-
-    return vehicle
-
-
-# ----------------------------------------------------------------------
-#   Define the Configurations
-# ---------------------------------------------------------------------
-
-def configs_setup(vehicle):
-    
-    # ------------------------------------------------------------------
-    #   Initialize Configurations
-    # ------------------------------------------------------------------
-    
-    configs = SUAVE.Components.Configs.Config.Container()
-    
-    base_config = SUAVE.Components.Configs.Config(vehicle)
-    base_config.tag = 'base'
-    configs.append(base_config)
-    
-    # ------------------------------------------------------------------
-    #   Cruise Configuration
-    # ------------------------------------------------------------------
-    
-    config = SUAVE.Components.Configs.Config(base_config)
-    config.tag = 'cruise'
-    
-    configs.append(config)
-    
-    
-    # ------------------------------------------------------------------
-    #   Takeoff Configuration
-    # ------------------------------------------------------------------
-    
-    config = SUAVE.Components.Configs.Config(base_config)
-    config.tag = 'takeoff'
-    
-    config.wings['main_wing'].flaps.angle = 20. * Units.deg
-    config.wings['main_wing'].slats.angle = 25. * Units.deg
-    
-    config.V2_VS_ratio = 1.21
-    config.maximum_lift_coefficient = 2.
-    
-    configs.append(config)
-    
-    
-    # ------------------------------------------------------------------
-    #   Landing Configuration
-    # ------------------------------------------------------------------
-
-    config = SUAVE.Components.Configs.Config(base_config)
-    config.tag = 'landing'
-    
-    config.wings['main_wing'].flaps_angle = 30. * Units.deg
-    config.wings['main_wing'].slats_angle = 25. * Units.deg
-
-    config.Vref_VS_ratio = 1.23
-    config.maximum_lift_coefficient = 2.
-    
-    configs.append(config)
-    
-    
-    # done!
-    return configs
-
-# ----------------------------------------------------------------------
-#   Sizing for the Vehicle Configs
-# ----------------------------------------------------------------------
-
-# ----------------------------------------------------------------------
-#   Plot Mission
-# ----------------------------------------------------------------------
-
-def plot_mission(results,line_style='bo-'):
-
-    # ------------------------------------------------------------------
-    #   Throttle
-    # ------------------------------------------------------------------
-    plt.figure("Throttle History")
-    axes = plt.gca()
-    for i in range(len(results.segments)):
-        time = results.segments[i].conditions.frames.inertial.time[:,0] / Units.min
-        eta  = results.segments[i].conditions.propulsion.throttle[:,0]
-        axes.plot(time, eta, line_style)
-    axes.set_xlabel('Time (mins)')
-    axes.set_ylabel('Throttle')
-    axes.grid(True)
-
-
-    # ------------------------------------------------------------------
-    #   Angle of Attack
-    # ------------------------------------------------------------------
-
-    plt.figure("Angle of Attack History")
-    axes = plt.gca()
-    for i in range(len(results.segments)):
-        time = results.segments[i].conditions.frames.inertial.time[:,0] / Units.min
-        aoa = results.segments[i].conditions.aerodynamics.angle_of_attack[:,0] / Units.deg
-        axes.plot(time, aoa, line_style)
-    axes.set_xlabel('Time (mins)')
-    axes.set_ylabel('Angle of Attack (deg)')
-    axes.grid(True)
-
-
-    # ------------------------------------------------------------------
-    #   Fuel Burn Rate
-    # ------------------------------------------------------------------
-    plt.figure("Fuel Burn Rate")
-    axes = plt.gca()
-    for i in range(len(results.segments)):
-        time = results.segments[i].conditions.frames.inertial.time[:,0] / Units.min
-        mdot = results.segments[i].conditions.weights.vehicle_mass_rate[:,0]
-        axes.plot(time, mdot, line_style)
-    axes.set_xlabel('Time (mins)')
-    axes.set_ylabel('Fuel Burn Rate (kg/s)')
-    axes.grid(True)
-
-
-    # ------------------------------------------------------------------
-    #   Altitude
-    # ------------------------------------------------------------------
-    plt.figure("Altitude")
-    axes = plt.gca()
-    for i in range(len(results.segments)):
-        time     = results.segments[i].conditions.frames.inertial.time[:,0] / Units.min
-        distance = results.segments[i].conditions.frames.inertial.position_vector[:,0] / Units.nautical_miles
-        altitude = results.segments[i].conditions.freestream.altitude[:,0] / Units.km
-        axes.plot(distance, altitude, line_style)
-    axes.set_xlabel('Distance (nm)')
-    axes.set_ylabel('Altitude (km)')
-    axes.grid(True)
-
-
-    # ------------------------------------------------------------------
-    #   Vehicle Mass
-    # ------------------------------------------------------------------
-    plt.figure("Vehicle Mass")
-    axes = plt.gca()
-    for i in range(len(results.segments)):
-        time = results.segments[i].conditions.frames.inertial.time[:,0] / Units.min
-        mass = results.segments[i].conditions.weights.total_mass[:,0]
-        axes.plot(time, mass, line_style)
-    axes.set_xlabel('Time (mins)')
-    axes.set_ylabel('Vehicle Mass (kg)')
-    axes.grid(True)
-
-
-    # ------------------------------------------------------------------
-    #   Aerodynamics
-    # ------------------------------------------------------------------
-    fig = plt.figure("Aerodynamic Forces")
-    for segment in results.segments.values():
-
-        time   = segment.conditions.frames.inertial.time[:,0] / Units.min
-        Lift   = -segment.conditions.frames.wind.lift_force_vector[:,2]
-        Drag   = -segment.conditions.frames.wind.drag_force_vector[:,0]
-        Thrust = segment.conditions.frames.body.thrust_force_vector[:,0]
-
-        axes = fig.add_subplot(4,1,1)
-        axes.plot( time , Lift , line_style )
-        axes.set_xlabel('Time (min)')
-        axes.set_ylabel('Lift (N)')
-        axes.grid(True)
-
-        axes = fig.add_subplot(4,1,2)
-        axes.plot( time , Drag , line_style )
-        axes.set_xlabel('Time (min)')
-        axes.set_ylabel('Drag (N)')
-        axes.grid(True)
-
-        axes = fig.add_subplot(4,1,3)
-        axes.plot( time , Thrust , line_style )
-        axes.set_xlabel('Time (min)')
-        axes.set_ylabel('Thrust (N)')
-        axes.grid(True)
-
-        try:
-            Pitching_moment = segment.conditions.stability.static.cm_alpha[:,0]
-            axes = fig.add_subplot(4,1,4)
-            axes.plot( time , Pitching_moment , line_style )
-            axes.set_xlabel('Time (min)')
-            axes.set_ylabel('Pitching_moment (~)')
-            axes.grid(True)            
-        except:
-            pass 
-
-    # ------------------------------------------------------------------
-    #   Aerodynamics 2
-    # ------------------------------------------------------------------
-    fig = plt.figure("Aerodynamic Coefficients")
-    for segment in results.segments.values():
-
-        time   = segment.conditions.frames.inertial.time[:,0] / Units.min
-        CLift  = segment.conditions.aerodynamics.lift_coefficient[:,0]
-        CDrag  = segment.conditions.aerodynamics.drag_coefficient[:,0]
-        Drag   = -segment.conditions.frames.wind.drag_force_vector[:,0]
-        Thrust = segment.conditions.frames.body.thrust_force_vector[:,0]
-
-        axes = fig.add_subplot(3,1,1)
-        axes.plot( time , CLift , line_style )
-        axes.set_xlabel('Time (min)')
-        axes.set_ylabel('CL')
-        axes.grid(True)
-
-        axes = fig.add_subplot(3,1,2)
-        axes.plot( time , CDrag , line_style )
-        axes.set_xlabel('Time (min)')
-        axes.set_ylabel('CD')
-        axes.grid(True)
-
-        axes = fig.add_subplot(3,1,3)
-        axes.plot( time , Drag   , line_style )
-        axes.plot( time , Thrust , 'ro-' )
-        axes.set_xlabel('Time (min)')
-        axes.set_ylabel('Drag and Thrust (N)')
-        axes.grid(True)
-
-
-    # ------------------------------------------------------------------
-    #   Aerodynamics 2
-    # ------------------------------------------------------------------
-    fig = plt.figure("Drag Components")
-    axes = plt.gca()
-    for i, segment in enumerate(results.segments.values()):
-
-        time   = segment.conditions.frames.inertial.time[:,0] / Units.min
-        drag_breakdown = segment.conditions.aerodynamics.drag_breakdown
-        cdp = drag_breakdown.parasite.total[:,0]
-        cdi = drag_breakdown.induced.total[:,0]
-        cdc = drag_breakdown.compressible.total[:,0]
-        cdm = drag_breakdown.miscellaneous.total[:,0]
-        cd  = drag_breakdown.total[:,0]
-
-        if line_style == 'bo-':
-            axes.plot( time , cdp , 'ko-', label='CD_P' )
-            axes.plot( time , cdi , 'bo-', label='CD_I' )
-            axes.plot( time , cdc , 'go-', label='CD_C' )
-            axes.plot( time , cdm , 'yo-', label='CD_M' )
-            axes.plot( time , cd  , 'ro-', label='CD'   )
-            if i == 0:
-                axes.legend(loc='upper center')            
-        else:
-            axes.plot( time , cdp , line_style )
-            axes.plot( time , cdi , line_style )
-            axes.plot( time , cdc , line_style )
-            axes.plot( time , cdm , line_style )
-            axes.plot( time , cd  , line_style )            
-
-    axes.set_xlabel('Time (min)')
-    axes.set_ylabel('CD')
-    axes.grid(True)
-
-    return
-
-def simple_sizing(configs):
-    
-    base = configs.base
-    base.pull_base()
-    
-    # zero fuel weight
-    base.mass_properties.max_zero_fuel = 0.9 * base.mass_properties.max_takeoff 
-    
-    # wing areas
-    for wing in base.wings:
-        wing.areas.wetted   = 2.0 * wing.areas.reference
-        wing.areas.exposed  = 0.8 * wing.areas.wetted
-        wing.areas.affected = 0.6 * wing.areas.wetted
-    
-    # fuselage seats
-    base.fuselages['fuselage'].number_coach_seats = base.passengers
-    
-    # diff the new data
-    base.store_diff()
-    
-    # ------------------------------------------------------------------
-    #   Landing Configuration
-    # ------------------------------------------------------------------
-    landing = configs.landing
-    
-    # make sure base data is current
-    landing.pull_base()
-    
-    # landing weight
-    landing.mass_properties.landing = 0.85 * base.mass_properties.takeoff
-    
-    # diff the new data
-    landing.store_diff()
-    
-    # done!
-    return
-
-# ----------------------------------------------------------------------
-#   Define the Mission
-# ----------------------------------------------------------------------
-    
-def mission_setup(analyses):
-    
-    # ------------------------------------------------------------------
-    #   Initialize the Mission
-    # ------------------------------------------------------------------
-    
-    mission = SUAVE.Analyses.Mission.Sequential_Segments()
-    mission.tag = 'the_mission'
-    
-    #airport
-    airport = SUAVE.Attributes.Airports.Airport()
-    airport.altitude   =  0.0  * Units.ft
-    airport.delta_isa  =  0.0
-    airport.atmosphere = SUAVE.Attributes.Atmospheres.Earth.US_Standard_1976()
-    
-    mission.airport = airport    
-    
-    # unpack Segments module
-    Segments = SUAVE.Analyses.Mission.Segments
-    
-    # base segment
-    base_segment = Segments.Segment()
-    
-    
-    # ------------------------------------------------------------------
-    #   First Climb Segment: constant Mach, constant segment angle 
-    # ------------------------------------------------------------------
-    
-    segment = Segments.Climb.Constant_Speed_Constant_Rate(base_segment)
-    segment.tag = "climb_1"
-    
-    segment.analyses.extend( analyses.takeoff )
-    
-    segment.altitude_start = 0.0   * Units.km
-    segment.altitude_end   = 3.0   * Units.km
-    segment.air_speed      = 125.0 * Units['m/s']
-    segment.climb_rate     = 6.0   * Units['m/s']
-    
-    # add to misison
-    mission.append_segment(segment)
-    
-    
-    # ------------------------------------------------------------------
-    #   Second Climb Segment: constant Speed, constant segment angle 
-    # ------------------------------------------------------------------    
-    
-    segment = Segments.Climb.Constant_Speed_Constant_Rate(base_segment)
-    segment.tag = "climb_2"
-    
-    segment.analyses.extend( analyses.cruise )
-    
-    segment.altitude_end   = 8.0   * Units.km
-    segment.air_speed      = 190.0 * Units['m/s']
-    segment.climb_rate     = 6.0   * Units['m/s']
-    
-    # add to mission
-    mission.append_segment(segment)
-    
-    
-    # ------------------------------------------------------------------
-    #   Third Climb Segment: constant Mach, constant segment angle 
-    # ------------------------------------------------------------------    
-    
-    segment = Segments.Climb.Constant_Speed_Constant_Rate(base_segment)
-    segment.tag = "climb_3"
-    
-    segment.analyses.extend( analyses.cruise )
-    
-    segment.altitude_end = 10.668 * Units.km
-    segment.air_speed    = 226.0  * Units['m/s']
-    segment.climb_rate   = 3.0    * Units['m/s']
-    
-    # add to mission
-    mission.append_segment(segment)
-    
-    
-    # ------------------------------------------------------------------    
-    #   Cruise Segment: constant speed, constant altitude
-    # ------------------------------------------------------------------    
-    
-    segment = Segments.Cruise.Constant_Speed_Constant_Altitude(base_segment)
-    segment.tag = "cruise"
-    
-    segment.analyses.extend( analyses.cruise )
-    
-    segment.air_speed  = 230.412 * Units['m/s']
-    segment.distance   = (3933.65 + 770) * Units.km
-        
-    mission.append_segment(segment)
-    
-    
-    # ------------------------------------------------------------------    
-    #   First Descent Segment: Linear Mach, constant rate segment
-    # ------------------------------------------------------------------    
-    
-    segment = Segments.Descent.Linear_Mach_Constant_Rate(base_segment)
-    segment.tag = "descent_1"
-    
-    segment.analyses.extend( analyses.cruise )
-    
-    segment.altitude_end = 5.0   * Units.km
-    #segment.air_speed    = 170.0 * Units['m/s']
-    segment.descent_rate = 5.0   * Units['m/s']
-    segment.mach_end    = 0.65
-    segment.mach_start  = 0.78    
-    
-    # add to mission
-    mission.append_segment(segment)
-    
-    
-    # ------------------------------------------------------------------    
-    #   Second Descent Segment: constant speed, constant segment rate
-    # ------------------------------------------------------------------    
-
-    segment = Segments.Descent.Constant_Speed_Constant_Rate(base_segment)
-    segment.tag = "descent_2"
-
-    segment.analyses.extend( analyses.landing )
-    
-    segment.altitude_end = 0.0   * Units.km
-    segment.air_speed    = 145.0 * Units['m/s']
-    segment.descent_rate = 5.0   * Units['m/s']
-
-    # append to mission
-    mission.append_segment(segment)
-    
-    # ------------------------------------------------------------------    
-    #   Mission definition complete    
-    # ------------------------------------------------------------------
-    
-    return mission
-
-def missions_setup(base_mission):
-
-    # the mission container
-    missions = SUAVE.Analyses.Mission.Mission.Container()
-    
-    # ------------------------------------------------------------------
-    #   Base Mission
-    # ------------------------------------------------------------------
-    
-    missions.base = base_mission
-    
-    
-    # ------------------------------------------------------------------
-    #   Mission for Constrained Fuel
-    # ------------------------------------------------------------------    
-    fuel_mission = SUAVE.Analyses.Mission.Mission() #Fuel_Constrained()
-    fuel_mission.tag = 'fuel'
-    fuel_mission.range   = 1277. * Units.nautical_mile
-    fuel_mission.payload   = 19000.
-    missions.append(fuel_mission)    
-    
-    
-    # ------------------------------------------------------------------
-    #   Mission for Constrained Short Field
-    # ------------------------------------------------------------------    
-    short_field = SUAVE.Analyses.Mission.Mission(base_mission) #Short_Field_Constrained()
-    short_field.tag = 'short_field'    
-    
-    #airport
-    airport = SUAVE.Attributes.Airports.Airport()
-    airport.altitude   =  0.0  * Units.ft
-    airport.delta_isa  =  0.0
-    airport.atmosphere = SUAVE.Attributes.Atmospheres.Earth.US_Standard_1976()
-    airport.available_tofl = 1500.
-    short_field.airport = airport    
-    missions.append(short_field)
-   
-
-    
-    # ------------------------------------------------------------------
-    #   Mission for Fixed Payload
-    # ------------------------------------------------------------------    
-    payload = SUAVE.Analyses.Mission.Mission() #Payload_Constrained()
-    payload.tag = 'payload'
-    payload.range   = 2316. * Units.nautical_mile
-    payload.payload   = 19000.
-    missions.append(payload)
-    
-    
-    # done!
-    return missions  
-
-def check_results(new_results,old_results):
-    
-    # check segment values
-    check_list = [
-        'segments.cruise.conditions.aerodynamics.angle_of_attack',
-        'segments.cruise.conditions.aerodynamics.drag_coefficient',
-        'segments.cruise.conditions.aerodynamics.lift_coefficient',
-        'segments.cruise.conditions.stability.static.cm_alpha',
-        'segments.cruise.conditions.stability.static.cn_beta',
-        'segments.cruise.conditions.propulsion.throttle',
-        'segments.cruise.conditions.weights.vehicle_mass_rate',
-    ]
-    
-    # do the check
-    for k in check_list:
-        print k
-        
-        old_val = np.max( old_results.deep_get(k) )
-        new_val = np.max( new_results.deep_get(k) )
-        err = (new_val-old_val)/old_val
-        print 'Error at Max:' , err
-        assert np.abs(err) < 1e-6 , 'Max Check Failed : %s' % k
-        
-        old_val = np.min( old_results.deep_get(k) )
-        new_val = np.min( new_results.deep_get(k) )
-        err = (new_val-old_val)/old_val
-        print 'Error at Min:' , err
-        assert np.abs(err) < 1e-6 , 'Min Check Failed : %s' % k        
-        
-        print ''
-    
-    ## check high level outputs
-    #def check_vals(a,b):
-        #if isinstance(a,Data):
-            #for k in a.keys():
-                #err = check_vals(a[k],b[k])
-                #if err is None: continue
-                #print 'outputs' , k
-                #print 'Error:' , err
-                #print ''
-                #assert np.abs(err) < 1e-6 , 'Outputs Check Failed : %s' % k  
-        #else:
-            #return (a-b)/a
-
-    ## do the check
-    #check_vals(old_results.output,new_results.output)
-    
-
-    return
-
-    
-def load_results():
-    return SUAVE.Input_Output.SUAVE.load('results_mission_B737.res')
-    
-def save_results(results):
-    SUAVE.Input_Output.SUAVE.archive(results,'results_mission_B737.res')
-    return
-    
-if __name__ == '__main__': 
-    main()    
-    plt.show()
-
